requirements.txt
flask
python-dotenv
Werkzeug
psycopg2
PyJWT
uuid
Flask-WTF
Flask-Cors